requirement.txt
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
Authlib==0.15.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cmake==3.20.3
colorama==0.4.4
contourpy==1.0.5
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
Django==3.1.6
django-crispy-forms==1.11.2
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
face-recognition-models==0.3.0
fastjsonschema==2.16.2
filelock==3.0.12
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
fonttools==4.37.4
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==6.6.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
mypy==0.910
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.2.54
openpyxl==3.0.10
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.0.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.1
psutil==5.9.2
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytube==11.0.0
pytz==2021.3
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.12
sqlparse==0.4.1
stack-data==0.5.1
terminado==0.16.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
traitlets==5.4.0
typing-extensions==4.0.1
uritemplate==4.1.1
url-for-s3==0.1.2
urllib3==1.26.7
uvicorn==0.17.6
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==4.0.3
WTForm==1.0
WTForms==3.0.1
zipp==3.15.0
zoom==0.0.15
